#21 - feat: 맛Place 평점 평균, Post 갯수 매핑 맛Post와 Join하여 star 정보를 받아와 평균을 구하고, Post 갯수를 구해 응답값에 넣어 전달

+++ server/matp/src/main/java/com/matp/place/entity/Place.java
@@ -4,9 +4,7 @@ package com.matp.place.entity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.annotation.*;
 import org.springframework.data.relational.core.mapping.Table;
 import java.time.LocalDateTime;
 
@@ -16,13 +14,14 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class Place {
     @Id
-    private int id;
+    private Long id;
 
     private String tel;
 
     private String address;
 
     private String roadNameAddress;
+
     private String name;
 
     private String point;
@@ -36,5 +35,4 @@ public class Place {
 
     @LastModifiedDate
     private LocalDateTime modifiedAt;
-
 }
